連続メッセージ送信のバグ修正等
